Responsive Tester
This is an aplicative totally Front-End, since your criation and aplication. The Idea is facilitate and speed up the responsive contruction, it's too easy to configurate and just need to be allocated in the main directory of your project.

Responsive Tester still in progress, and is opened for understanding of the Code as helps to new implementations, fell it confortable.